Rename files (remove middleware suffix)
