+++ src/index.ts
@@ -1,3 +1,3 @@
-export * from "./middlewares/build-model.middleware";
-export * from "./middlewares/build-where-query.middleware";
+export * from "./middlewares/build-model";
+export * from "./middlewares/build-where-query";
 export * from "./middlewares/order-model";
